requirements.txt
sqlalchemy>=2.0
psycopg2-binary>=2.9
pandas>=2.2
python-dotenv>=1.0
ipykernel>=6.29
jupyter>=1.0
